latest

+++ extractExtremeSubjects.20210602.sh
@@ -8,7 +8,7 @@ gene=$1
 
 
 # exomesFile=/SAN/ugi/UGIbiobank/data/downloaded/ukb41465.exomes.txt
-model=LOF.20201227
+model=LOF.20210615
 argFile=~/pars/gva.UKBB.$model.arg
 exomesFile=/SAN/ugi/UGIbiobank/data/downloaded/ukb41465.exomes.20201103.txt
 
@@ -30,8 +30,8 @@ if ($17 != "" && $17!="comment")
  split(l[2],p,"-");
  split(p[1],pp,"."); 
  split(p[2],a,","); 
- if ($2>$6)  print l[1] ":" pp[1],a[1],a[2],$1,"outputAlt"
- else print l[1] ":" pp[1],a[1],a[2],$1,"outputRef"
+ if ($2>$6)  print l[1] ":" pp[1],a[1],a[2],$1,"outputAlt",$17
+ else print l[1] ":" pp[1],a[1],a[2],$1,"outputRef",$17
  } 
 }' > $model.$gene.vars.txt
 # with split loci may be . in position
@@ -43,7 +43,7 @@ if [ ! -s $model.$gene.vars.txt ]
 then
   exit
 fi
-cat $model.$gene.vars.txt | while read a b c d e
+cat $model.$gene.vars.txt | while read a b c d e f
 do
   if [ ! -e altSubs.$model.$gene.$a.$b.$c.txt ]
   then
@@ -54,7 +54,7 @@ do
       showAltSubs --arg-file $argFile --position $a --ref-all $b --alt-all $c --test-name altSubs.$model.$gene.$a.$b.$c
 	fi
   fi
-  cat altSubs.$model.$gene.$a.$b.$c.txt | awk -v v=$d '{ print $1, v }' >> IDsAndAnnots.$model.$gene.txt
+  cat altSubs.$model.$gene.$a.$b.$c.txt | awk -v v=$d -v w=$f '{ print $1, v, w }' >> IDsAndAnnots.$model.$gene.txt
   cat altSubs.$model.$gene.$a.$b.$c.txt | while read s rest
   do
    echo $s >> IDs.$model.$gene.txt
@@ -63,12 +63,34 @@ do
 done 
 cat IDsAndAnnots.$model.$gene.txt | sort > IDsAndAnnots.sorted.$model.$gene.txt
 cat IDs.$model.$gene.txt  | sort > IDs.sorted.$model.$gene.txt
+if [ ! -e $model.$gene.phenos.txt ]
+then
 echo eid$'\t'extra > extraField.txt
 head -n 1 $exomesFile | join -t  $'\t' - extraField.txt > $model.$gene.phenos.txt
 # this is because the exomes file has got a tab appended to each line after the header
 # and otherwise fread will not work
 tail -n +2 $exomesFile | join -t  $'\t' IDs.sorted.$model.$gene.txt - >> $model.$gene.phenos.txt 
 # grep -f  IDs.sorted.$model.$gene.txt $exomesFile >> $model.$gene.phenos.txt # this is not really safe
+fi
+
+if [ ! -e IDsAndCounts.$model.$gene.txt ]
+then
+  cat IDsAndAnnots.sorted.$model.$gene.txt | while read ID annot fullAnnot
+  do
+    IFS='-', read chrPos rest <<<$annot
+	IFS=':', read chr pos <<<$chrPos
+	IFS=',', read all1 rest <<<$rest
+	IFS=':', read all2 rest <<< $rest
+	if [ ${#all1} -gt ${#all2} ] # deletion
+	then
+	  newpos=$((pos + 1))
+	else
+	  newpos=$pos
+	fi
+	bash /home/rejudcu/UKBB/UKBBscripts/viewReads.sh $ID $chr $newpos
+	echo $ID$'\t'$annot$'\t'`cat /home/rejudcu/UKBB/LOF/snapShots/$ID.$chr.$newpos.countsByStrand.txt`$'\t'$fullAnnot >> IDsAndCounts.$model.$gene.txt
+  done
+fi
 
 # then R can read in the two files and get the column headings right
 # can use IDs file to get a phenotype of who has any LOF variants for other analyses
